+++ Basic.WebApi/DTOs/EventForView.cs
@@ -46,20 +46,20 @@ namespace Basic.WebApi.DTOs
         /// </summary>
         [Required]
         [SwaggerSchema(Format = "number/hours")]
-        public int DurationFirstDay { get; set; }
+        public decimal DurationFirstDay { get; set; }
 
         /// <summary>
         /// Gets or sets the number of hours associated to the last day.
         /// </summary>
         [Required]
         [SwaggerSchema(Format = "number/hours")]
-        public int DurationLastDay { get; set; }
+        public decimal DurationLastDay { get; set; }
 
         /// <summary>
         /// Gets or sets the total duration of the event, in hours.
         /// </summary>
         [Required]
         [SwaggerSchema(Format = "number/hours")]
-        public int DurationTotal { get; set; }
+        public decimal DurationTotal { get; set; }
     }
 }
